+++ src/model/Map.java
@@ -1,16 +1,19 @@
 package model;
 
+import java.util.Timer;
+import java.util.TimerTask;
+
 public class Map{
     private Cell mappa[][];
     public final Robot robot;
+    private Timer timer;
 
     protected Map(Cell[][] mappa){
         this.mappa = mappa;
         this.robot = new Robot(this);
     }        
 
-    public Cell getCasella(int i, int j)
-    {
+    public Cell getCasella(int i, int j){
         return this.mappa[i][j];
     }
 
@@ -22,12 +25,25 @@ public class Map{
         return this.mappa[0].length;
     }
 
-    protected void setNewRobotPosition(int oldI, int oldJ, int newI, int newJ)
-    {
+    protected void setNewRobotPosition(int oldI, int oldJ, int newI, int newJ){
         //TODO Need fixing
         this.mappa[oldI][oldJ] = null;//new Vuoto(oldI, oldJ);
     }
 
-    public void event()
-    {}
+    public void event(Map mappa){
+        
+    }
+
+    public void startTimer(long delay, long period){
+        this.timer = new Timer();
+        timer.scheduleAtFixedRate(new TimerTask() {
+            public void run() {
+                event(Map.this);
+            }
+        }, delay, period);
+    }
+
+    public void stopTimer(){
+        timer.cancel();
+    }
 }
